requirements/essentials.txt
# Python
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
requests==2.26.0
urllib3==1.26.6
six==1.16.0
nanoid==2.0.0
pygeoip==0.3.2

# django
Django==3.2.6
asgiref==3.4.1
pytz==2022.1
sqlparse==0.4.1

# .env Management
python-dotenv==0.19.2 # read .env file
python-decouple==3.6 # TypeCast Env Vars

# Django Settings Management
django-configurations==2.3.2 #class-based Django settings
dj-database-url==0.5.0 # DB URLs
django-constance==2.8.0 # Dynamic Django settings
django-picklefield==3.0.1 # database backend for constance

# django REST framework
djangorestframework==3.12.4
djangorestframework-gis==0.18 # for POST-Gis

# for Authentication
djangorestframework-simplejwt==5.1.0

# for parsing
djangorestframework_camel_case==1.3.0

# Django Filters - addtional retrive query lookups
django-filter==21.1

# Datetime Manipulation - To be Used
pendulum==2.1.2
python-dateutil==2.8.2
pytzdata==2020.1

# used in infinity-base
Markdown==3.3.4

# Redis - Caching, DB for infinite-redis
redis==3.5.3
django-redis==5.0.0
django-redisboard==8.3.0

# DB Adapter for infinity-psql
psycopg2-binary==2.9.3 # PostgreSQL
django-postgres-extra==2.0.5 # PSQL Extra
django-timescaledb==0.2.11 # Timescale DB

# Django Models
django-polymorphic==3.1.0 # Easier Inheritance
django-simple-history==3.1.1 # stores model state on every change
django-model-utils==4.2.0

# Tenant
django-tenants==3.4.2
tenant-schemas-celery==2.0.0

# Celery[redis]
celery==5.2.6 #
django_celery_beat==2.2.1
django-celery-results==2.3.1
amqp==5.1.1
billiard==3.6.4.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
click==8.1.3
kombu==5.2.4
vine==5.0.0
django-timezone-field==4.2.3
python-crontab==2.6.0

# Django Channels (channels_redis[cryptography])
Automat==20.2.0
aioredis==1.3.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
constantly==15.1.0
cryptography==36.0.1
hiredis==2.0.0
hyperlink==21.0.0
incremental==21.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyopenssl==22.0.0
service-identity==21.1.0
twisted==22.2.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
zope.interface==5.4.0

# graphene
aniso8601-7.0.0
graphene==2.1.9
graphene_django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1

# Django Admin Actions
django-adminactions==1.15
xlrd==2.0.1
xlwt==1.3.0
